pyproject.toml
[project]
name = "lambda-flask"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.36.5",
    "emoji>=2.14.1",
    "flask-openapi3==2.5.5",
    "fuzzywuzzy>=0.18.0",
    "google-api-python-client>=2.159.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.1",
    "pre-commit>=4.1.0",
    "psycopg2>=2.9.10",
    "pyairtable==2.3.7",
    "pydantic-core>=2.27.2",
    "python-dotenv>=1.0.1",
    "python-levenshtein>=0.26.1",
    "requests>=2.32.3",
    "ruff>=0.9.4",
    "sqlalchemy>=2.0.37",
    "sqlalchemy-utils>=0.41.2",
]
